move cell output element to its own module
